% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dur_ita_pol}
\alias{dur_ita_pol}
\title{Acousics and articulatory durational measures of Italian and Polish}
\format{
A tibble with 1343 observations and 52 variables.
\describe{
\item{\code{index}}{observation number within speaker}
\item{\code{speaker}}{speaker's ID}
\item{\code{file}}{audio chunk file name}
\item{\code{rec_date}}{date and time of recording}
\item{\code{ipu}}{SPPAS IPU index}
\item{\code{prompt}}{sentence stimulys}
\item{\code{word}}{word stimulus}
\item{\code{time}}{(s) time of the sentence onset within the concatenated audio file}
\item{\code{sentence_ons}}{(s) onset time of the sentence}
\item{\code{sentence_off}}{(s) offset time of the sentence}
\item{\code{word_ons}}{(s) onset time of the target word}
\item{\code{word_off}}{(s) offsett time of the target word (= C1 onset)}
\item{\code{v1_ons}}{(s) onset time of V1 (= C1 offset)}
\item{\code{c2_ons}}{(s) onset time of C2 (= V1 offset)}
\item{\code{v2_ons}}{(s) onset time of V2 (= C2 offset)}
\item{\code{c1_rel}}{(s) time of C1 release}
\item{\code{c2_rel}}{(s) time of C2 release}
\item{\code{voicing_start}}{(s) time of voicing onset}
\item{\code{voicing_end}}{(s) time of voicing offset}
\item{\code{voicing_duration}}{(ms) duration of voiced interval}
\item{\code{voiced_points}}{number of points out of 5 within the first half of C1 closure in which voicing is present}
\item{\code{GONS}}{(s) onset of C1 closing gesture}
\item{\code{max}}{(s) time of maximum displacement of C1 closing gesture}
\item{\code{NOFF}}{(s) offset of C1 gesture nucleus}
\item{\code{NONS}}{(s) onset of C1 gesture nucleus}
\item{\code{peak1}}{(s) first tongue velocity peak}
\item{\code{peak2}}{(s) second tongue velocity peak}
\item{\code{c1_duration}}{(ms) duration of C1}
\item{\code{c1_clos_duration}}{(ms) duration of C1 closure}
\item{\code{c1_vot}}{(ms) C1 Voice Onset Time}
\item{\code{c1_rvofft}}{(ms) C1 release to V1 offset time}
\item{\code{v1_duration}}{(ms) duration of V1}
\item{\code{c1_duration}}{(ms) duration of C1}
\item{\code{c2_clos_duration}}{(ms) duration of C2 closure}
\item{\code{v2_duration}}{(ms) duration of V2}
\item{\code{v_v}}{(ms) V1 onset to V2 onset (Vowel-to-Vowel) duration}
\item{\code{word_duration}}{(ms) duration of the word}
\item{\code{sentence_duration}}{(s) duration of sentence}
\item{\code{language}}{speaker's native language (\code{Italian}, \code{Polish})}
\item{\code{gender}}{speaker's sex (\code{f}, \code{m})}
\item{\code{glottocode}}{language Glottocode}
\item{\code{item}}{word ID number}
\item{\code{ipa}}{IPA transcription of the word}
\item{\code{c1}}{first consonant (C1)}
\item{\code{c1_phonation}}{voicing of C1 (\code{voiceless}, \code{voiced})}
\item{\code{vowel}}{V1 and V2 (\code{a}, \code{o}, \code{u})}
\item{\code{anteropost}}{backness of the vowel (\code{back}, \code{central})}
\item{\code{height}}{height of the vowel (\code{high}, \code{mid}, \code{low})}
\item{\code{c2}}{second consonant (C2)}
\item{\code{c2_phonation}}{voicing of C2 (\code{voiceless} or \code{voiced})}
\item{\code{c2_place}}{place of C2 (\code{coronal}, \code{velar})}
\item{\code{speech_rate}}{speech rate as syllables per second}
\item{\code{speech_rate_c}}{centered speech rate as syllables per second}
}
}
\usage{
dur_ita_pol
}
\description{
A dataset with measures from acoustics and articulatory data of Italian and Polish.
The dataset contains durational measures of segments, voicing, and tongue gestures.
}
\keyword{datasets}
