% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{everywhere_loc}
\alias{everywhere_loc}
\title{'Everywhere here can say this': The English locative impersonal}
\format{
A \link[tibble:tibble]{tibble::tibble} with 5000 rows and 7 variables:
\describe{
\item{Subject}{Subject unique identifier.}
\item{Item}{Word.}
}
}
\source{
Sluckin, Benjamin L. and Itamar Kastner. 2022. 'Everywhere here can say this': The English locative impersonal. LingBuzz. https://ling.auf.net/lingbuzz/006708
}
\usage{
everywhere_loc
}
\description{
'Everywhere here can say this': The English locative impersonal
}
\keyword{datasets}
