% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft1_data}
\alias{ft1_data}
\title{FT1: Data visualisation}
\format{
A \link[tibble:tibble]{tibble::tibble} with 2530 rows and 7 variables:
\describe{
\item{Subject}{Subject identifier.}
\item{Condition}{Experimental condition (\code{gold-bad}, \code{gold-good}, \code{simple}, \code{adverb}, \code{modal}).}
\item{Restrictor}{Whether the sentence contained a restrictor or not (\code{non-restricted}, \code{restricted}).}
\item{Value}{Acceptability scale (Likert scale: 1-7).}
\item{Age}{Subject's age group.}
\item{Education}{Subject's education level.}
\item{Gender}{Subject's self-reported gender.}
}
}
\source{
Sluckin, Benjamin L. and Itamar Kastner. 2022. 'Everywhere here can say this': The English locative impersonal. LingBuzz. https://ling.auf.net/lingbuzz/006708
}
\usage{
ft1_data
}
\description{
A selection of data from Sluckin and Kastner 2022. Only UK participants
from four experimental conditions are included.
}
\keyword{datasets}
