% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gael_morph}
\alias{gael_morph}
\title{Morphology and dialectology in the Linguistic Survey of Scotland}
\format{
A \link[tibble:tibble]{tibble::tibble} with 201 rows and 74 variables:
\describe{
\item{sgds_point}{SGDS point.}
\item{id}{Unique identifier.}
}
}
\source{
Iosad, Pavel & William Lamb. 2020. \emph{Morphology and dialectology in
the Linguistic Survey of Scotland}. OSF.
\href{https://osf.io/4bp2y}{osf.io/4bp2y}.
}
\usage{
gael_morph
}
\description{
The \href{https://osf.io/4bp2y}{Morphology and dialectology in the Linguistic Survey of Scotland} contains a portion of the morphological
questionnaire returns collected for the Linguistic Survey of Scotland
(Gaelic). The data have been transcribed and coded from questionnaire slips
by Will Lamb.
}
\keyword{datasets}
