% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gestures}
\alias{gestures}
\title{A Cross-Cultural Analysis of Early Prelinguistic Gesture Development and Its Relationship to Language Development}
\format{
A \link[tibble:tibble]{tibble::tibble} with 5000 rows and 7 variables:
\describe{
\item{Subject}{Subject unique identifier.}
\item{Item}{Word.}
}
}
\source{
Thea Cameron-Faulkner, Nivedita Malik, Circle Steele, Stefano Coretta, Ludovica Serratrice, Elena Lieven. 2020. A cross cultural analysis of early prelinguistic gesture development and its relationship to language development. Child Development 92:1. 273-290. DOI: https://doi.org/10.1111/cdev.13406
}
\usage{
gestures
}
\description{
A Cross-Cultural Analysis of Early Prelinguistic Gesture Development and Its Relationship to Language Development
}
\keyword{datasets}
