% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mald_1_1}
\alias{mald_1_1}
\title{Massive Auditory Lexical Decision 1.1}
\format{
A \link[tibble:tibble]{tibble::tibble} with 5000 rows and 7 variables:
\describe{
\item{Subject}{Subject unique identifier.}
\item{Item}{Word.}
\item{IsWord}{Whether it is a real word or a nonce word.}
\item{PhonLev}{Mean phoneme-level Levenshtein distance.}
\item{RT}{Reaction times (ms).}
\item{ACC}{Accuracy of lexical decision.}
\item{RT_log}{Logged reaction times.}
}
}
\source{
Tucker, Benjamin V. 2021. Massive Auditory Lexical Decision 1.1. DOI: https://doi.org/10.7939/r3-kh0r-r116.
}
\usage{
mald_1_1
}
\description{
Massive Auditory Lexical Decision 1.1
}
\keyword{datasets}
