% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shallow}
\alias{shallow}
\title{Second Language Users exhibit shallow morphological processing}
\format{
A \link[tibble:tibble]{tibble::tibble} with 1950 rows and 11 variables:
\describe{
\item{Group}{Participant group (\code{L1}, \code{L2}).}
\item{ID}{Subject unique identifier.}
\item{List}{Word list.}
\item{Target}{Target word.}
\item{RT}{Reaction time (ms).}
\item{RT_log}{Logged reaction time.}
\item{Critical_Filler}{Whether the trial is a \code{critical} or a \code{filler} trial.}
\item{Word_Nonword}{Whether the word is a real \code{word} or a nonce word (only \code{word} is present in the data).}
\item{Relation_type}{Whether the relation type is \code{Unrelated}, \code{Constituent}, or \code{NonConstituent}.}
\item{Branching}{Whether the trial word is \code{Left}-branching or \code{Right}-branching.}
\item{accuracy}{Whether the trial was \code{correct} or \code{incorrect}.}
}
}
\source{
Song, Y., Do, Y., Thompson, A., Waegemaekers, E., & Lee, J. 2020.
SECOND LANGUAGE USERS EXHIBIT SHALLOW MORPHOLOGICAL PROCESSING. Studies in
Second Language Acquisition, 42(5), 1121-1136.
doi:10.1017/S0272263120000170
}
\usage{
shallow
}
\description{
Second Language Users exhibit shallow morphological processing
}
\keyword{datasets}
