% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqmf-package.R
\docType{package}
\name{sqmf-package}
\alias{sqmf}
\alias{sqmf-package}
\title{sqmf: Statistics and Quantitative Methods (Frequentist)}
\description{
Materials and slides for the SQM (Frequentist) course at UoE LEL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stefanocoretta/sqmf}
  \item \url{https://stefanocoretta.github.io/sqmf/}
  \item Report bugs at \url{https://github.com/stefanocoretta/sqmf/issues}
}

}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID}) (stefanocoretta)

}
\keyword{internal}
